Refactor 'and' and 'or' test packages to be prefixed with 'logical_'.
