+++ src/test/java/tests/expressions/logical_and/all-true.frendli
