+++ src/test/java/tests/expressions/logical_and/operand-updates-state-until-false.frendli
