+++ src/test/java/tests/expressions/logical_or/operand-updates-state-until-true.frendli
