+++ src/test/java/tests/expressions/logical_or/operand-updates-state.frendli
